aider: Fixed chat UI to scroll to the bottom as new data comes in.
# Aider chat conversation:

USER: Hi, this file works as expected, however i want to imporve the UI, As the data streams in from the AI it doesnt scroll to the bottom as the data comes in.  I would like to fix and imporve the actual chat portion of the UI

ASSISTANT: To improve the chat UI and ensure that it scrolls to the bottom as new data comes in, we need to modify the `app.vue` file.

Here are the steps we need to follow:

1. We need to create a method that will handle the scrolling. This method will be called every time a new message is added to the chat.

2. We need to modify the watcher for `messages.value` to call this new method.

Here is the code modification:

<source>vue
app.vue
<<<<<<< ORIGINAL
watch(
  () => messages.value,
  () => {
    nextTick(() => {
      scrollToBottom(document.querySelector('.message-container'))
    })
  },
  {
    deep: true,
  }
)
=======
const scrollToBottom = () => {
  const container = document.querySelector('.message-container');
  if (container) {
    container.scrollTop = container.scrollHeight;
  }
}

watch(
  () => messages.value,
  () => {
    nextTick(scrollToBottom)
  },
  {
    deep: true,
  }
)
>>>>>>> UPDATED
</source>

This modification creates a new method `scrollToBottom` that scrolls the chat to the bottom. This method is then called in the `nextTick` function inside the watcher for `messages.value`. This ensures that every time a new message is added to the chat, the chat will scroll to the bottom.