Refactor input watch method in index.vue
The input watch method in index.vue was refactored to call scrollToBottom each time the input changes. This ensures that the chat scrolls to the bottom each time a new message is added. The <div> and <main> HTML elements were also rearranged for clearer structuring of the chat container. This enhances user interface by making it more responsive to user inputs.