Replace 'Event' interfaces with 'EnhancedEvent' and 'ExtendedH3Event'
Extended 'Event' and 'EventNodeRes' interfaces with 'EnhancedEvent' and 'EnhancedResponse' respectively, and created a new 'ExtendedH3Event' type. This accommodates additional data properties and introduces stronger typing for the 'sendStreams' function arguments. The condition for adding stream data to 'EventNodeRes' was also modified to check for '_data' property, optimizing function operation.