Update UI dark theme and refactor components
This commit updates the dark theme color scheme for a number of components and refactors the navigation in the Sidebar component. It also modifies the configuration for the color mode in settings and index pages. Some adjustments have been made in the code related to 'h3' dependency and adjustments in package dependencies.